+++ src/myFP2ESP/FocuserSetupData.h
@@ -43,7 +43,7 @@ class SetupData
     byte get_backlash_out_enabled();
     byte get_tempcoefficient();
     byte get_tempprecision();
-    byte get_stepmode();
+    int  get_stepmode();
     byte get_coilpower();
     byte get_reversedirection();
     byte get_stepsizeenabled();
@@ -86,7 +86,7 @@ class SetupData
     void set_backlash_out_enabled(byte);
     void set_tempcoefficient(byte);
     void set_tempprecision(byte);
-    void set_stepmode(byte);
+    void set_stepmode(int);
     void set_coilpower(byte);
     void set_reversedirection(byte);
     void set_stepsizeenabled(byte);
@@ -154,7 +154,7 @@ class SetupData
     byte backlash_out_enabled;      // if 1, backlash is enabled for OUT movements (higher or +ve moves)
     byte tempcoefficient;           // steps per degree temperature coefficient value (maxval=255)
     byte tempprecision;             // 9 -12
-    byte stepmode;                  // stepping mode
+    int stepmode;                   // stepping mode
     byte coilpower;                 // if 1, coil power is enabled
     byte reversedirection;          // if 1, motor direction is reversed
     byte stepsizeenabled;           // if 1, controller returns step size
